requirements.txt
Flask
python-dotenv
reportlab
pandas
numpy
scikit-learn
